# Fix: INC FP8 quantization compatibility