### Summary
This PR addresses an issue where Intel Neural Compressor (INC) could not correctly apply FP8 quantization to the `GaudiMixtralSparseMoeBlock` module.

The problem was that the `hpu.mixture_of_experts` call was directly embedded within the forward method, making it difficult for INC to patch and replace for its quantization workflow.

### Changes
- Extracted the custom op call logic into a new method, `call_dynamic_moe_op`.
- Modified the forward method to call this new function.

This change allows INC to successfully patch the `call_dynamic_moe_op` method, enabling correct FP8 quantization of the `GaudiMixtralSparseMoeBlock` layer.

### Motivation
Previously,
```sh
# measure
>>> TQDM_DISABLE=1 PT_HPU_LAZY_MODE=1 QUANT_CONFIG=/root/optimum-habana/examples/text-generation/quantization_config/maxabs_measure.json  python3 /root/optimum-habana/examples/gaudi_spawn.py --use_deepspeed --world_size 2 /root/optimum-habana/examples/text-generation/run_generation.py --model_name_or_path mistralai/Mixtral-8x7B-Instruct-v0.1 --bf16 --use_hpu_graphs --limit_hpu_graphs --batch_size 1 --use_kv_cache --bucket_size 128 --max_new_tokens 128 --reuse_cache --bucket_internal

# quantize
>>> TQDM_DISABLE=1 PT_HPU_LAZY_MODE=1 QUANT_CONFIG=/root/optimum-habana/examples/text-generation/quantization_config/maxabs_quant_mixtral.json  python3 /root/optimum-habana/examples/gaudi_spawn.py --use_deepspeed --world_size 2 /root/optimum-habana/examples/text-generation/run_generation.py --model_name_or_path mistralai/Mixtral-8x7B-Instruct-v0.1 --bf16 --use_kv_cache --use_hpu_graphs --limit_hpu_graphs --batch_size 1 --max_new_tokens 1  --max_input_tokens 128  --n_iterations 1  --reuse_cache --bucket_size 128 --bucket_internal

File "/usr/local/lib/python3.12/dist-packages/neural_compressor/torch/algorithms/fp8_quant/_quant_common/helper_modules.py", line 1429, in forward
[rank0]:     raise Exception(
[rank0]: Exception: Error - Layer 'model.layers.0.block_sparse_moe' was called but was not quantized because no measures were supplied.
```

This PR
```sh
# quantize
>>> TQDM_DISABLE=1 PT_HPU_LAZY_MODE=1 QUANT_CONFIG=/root/optimum-habana/examples/text-generation/quantization_config/maxabs_quant_mixtral.json  python3 /root/optimum-habana/examples/gaudi_spawn.py --use_deepspeed --world_size 2 /root/optimum-habana/examples/text-generation/run_generation.py --model_name_or_path mistralai/Mixtral-8x7B-Instruct-v0.1 --bf16 --use_kv_cache --use_hpu_graphs --limit_hpu_graphs --batch_size 1 --max_new_tokens 1  --max_input_tokens 128  --n_iterations 1  --reuse_cache --bucket_size 128 --bucket_internal

Warming up iteration 3/3
09/04/2025 20:54:26 - INFO - __main__ - Time to first token = 24.576283001806587ms
09/04/2025 20:54:26 - INFO - __main__ - Time to rest of tokens = 0ms
09/04/2025 20:54:26 - INFO - __main__ - End to end latency = 24.915763991884887ms
09/04/2025 20:54:26 - INFO - __main__ - Running generate...
09/04/2025 20:54:26 - INFO - __main__ - Time to first token = 24.285959982080385ms
09/04/2025 20:54:26 - INFO - __main__ - Time to rest of tokens = 0ms
09/04/2025 20:54:26 - INFO - __main__ - End to end latency = 24.651726009324193ms
09/04/2025 20:54:26 - INFO - __main__ - Finished running generate

Input/outputs:
input 1: ('DeepSpeed is a machine learning framework',)
output 1.1: ('DeepSpeed is a machine learning framework that',)


Stats:
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Input tokens
Throughput (including tokenization) = 40.25759872238374 tokens/second
Average first token latency         = 24.285959982080385 ms
Average rest token latency          = 0.0 ms
Average end to end latency          = 24.651726009324193 ms
Memory allocated                    = 23.46 GB
Max memory allocated                = 44.2 GB
Total memory available              = 94.62 GB
Graph compilation duration          = 4.1490547669818625 seconds
```

### Impact
This refactoring has no impact on the functional behavior of the model. It solely addresses a compatibility issue with Intel Neural Compressor (INC). By moving the custom operator call into a dedicated function, we provide a stable hook that INC can patch for its FP8 quantization process. This change ensures that the model can be successfully quantized using INC without altering the model's performance or output.

## Before submitting
- [ ] This PR fixes a typo or improves the docs (you can dismiss the other checks if that's the case).
- [ ] Did you make sure to update the documentation with your changes?
- [ ] Did you write any new necessary tests?