@yafshar 2025-09-04
@astachowiczhabana I've addressed the FP8 quantization issue you raised. Please review! thanks